fix openresty brotli library missing
The currently installed Brotli is downloaded directly from GitHub; however, the deps directory included within it is empty. This means that additional steps are required to download the necessary dependencies separately. If these dependencies are not acquired, you will encounter an error stating "brotli is missing" during the image packaging process.